file system clean up and pages screen renameds
