--- src/app/[locale]/seller-registration.tsx
@@ -1,12 +1,0 @@
-import SellerRegistration from '@/components/seller/add-edit-seller/SellerRegistration';
-
-const SellerRegistrationPage: React.FC = () => {
-  return (
-    <div>
-      <h1>Seller Registration</h1>
-      <SellerRegistration />
-    </div>
-  );
-};
-
-export default SellerRegistrationPage;
