--- src/app/[locale]/seller-registration/page.tsx
@@ -1,9 +1,0 @@
-// app/seller-registration/page.tsx
-import React from 'react';
-import SellerRegistration from '@/components/seller/add-edit-seller/SellerRegistration'; // Adjust the import path as necessary
-
-const SellerRegistrationPage: React.FC = () => {
-  return <SellerRegistration />;
-};
-
-export default SellerRegistrationPage;
