--- src/components/seller/add-edit-seller/seller-register.module.css
@@ -1,105 +1,0 @@
-.container {
-    padding: 16px;
-  }
-  
-  .header {
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 16px;
-  }
-  
-  .header button {
-    background: none;
-    border: none;
-    cursor: pointer;
-  }
-  
-  .header button svg {
-    width: 24px;
-    height: 24px;
-  }
-  
-  .title {
-    font-size: 1.25rem;
-    font-weight: bold;
-    margin-bottom: 16px;
-  }
-  
-  .label {
-    display: block;
-    margin-bottom: 8px;
-    font-weight: bold;
-  }
-  
-  .input, .textarea, .select {
-    width: 100%;
-    padding: 8px;
-    margin-bottom: 16px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-  }
-  
-  .textarea {
-    height: 100px;
-  }
-  
-  .button {
-    display: inline-block;
-    padding: 10px 20px;
-    background-color: #4caf50;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    text-align: center;
-  }
-  
-  .button:disabled {
-    background-color: #ccc;
-    cursor: not-allowed;
-  }
-  
-  .reviewContainer {
-    margin-bottom: 16px;
-  }
-  
-  .reviewSummary {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin-bottom: 8px;
-  }
-  
-  .rangeContainer {
-    display: flex;
-    align-items: center;
-    width: 100%;
-  }
-  
-  .rangeContainer span {
-    margin: 0 8px;
-  }
-  
-  .uploadContainer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    border: 2px dashed #ccc;
-    padding: 20px;
-    margin-bottom: 16px;
-    text-align: center;
-  }
-  
-  .uploadContainer img {
-    width: 50px;
-    height: 50px;
-    margin-bottom: 8px;
-  }
-  
-  .uploadError {
-    color: red;
-    font-size: 0.875rem;
-    margin-top: 8px;
-  }
-
