--- src/components/shared/Review/form.tsx
@@ -1,7 +1,0 @@
-export default function ReviwForm(){
-    return (
-        <form>
-
-        </form>
-    )
-}
